moved hook to RunArea.js
