+++ frontend/components/RunArea.js
@@ -1,4 +1,4 @@
-import { html } from "../imports/Preact.js"
+import { html, useEffect, useState } from "../imports/Preact.js"
 
 export const RunArea = ({ runtime, onClick }) => {
     return html`
@@ -22,4 +22,29 @@ const prettytime = (time_ns) => {
     const roundedtime = time_ns.toFixed(time_ns >= 100.0 ? 0 : 1)
 
     return roundedtime + "\xa0" + prefices[i] + "s"
+}
+
+const update_interval = 50
+/**
+ * Returns the milliseconds passed since the argument became truthy.
+ * If argument is falsy, returns undefined.
+ *
+ * @param {boolean} truthy
+ */
+export const useMillisSinceTruthy = (truthy) => {
+    const [now, setNow] = useState(0)
+    const [startRunning, setStartRunning] = useState(0)
+    useEffect(() => {
+        let interval
+        if (truthy) {
+            const now = +new Date()
+            setStartRunning(now)
+            setNow(now)
+            interval = setInterval(() => setNow(+new Date()), update_interval)
+        }
+        return () => {
+            interval && clearInterval(interval)
+        }
+    }, [truthy])
+    return truthy ? now - startRunning : undefined
 }
