--- frontend/imports/hooks.js
@@ -1,26 +1,0 @@
-import { useEffect, useState } from "./Preact.js"
-
-const UPDATE_RUNNING_TIME = 50
-/**
- * Returns the milliseconds passed since the argument became truthy.
- * If argument is falsy, returns undefined.
- *
- * @param {boolean} truthy
- */
-export const useMillisSinceTruthy = (truthy) => {
-    const [now, setNow] = useState(0)
-    const [startRunning, setStartRunning] = useState(0)
-    useEffect(() => {
-        let interval
-        if (truthy) {
-            const now = +new Date()
-            setStartRunning(now)
-            setNow(now)
-            interval = setInterval(() => setNow(+new Date()), UPDATE_RUNNING_TIME)
-        }
-        return () => {
-            interval && clearInterval(interval)
-        }
-    }, [truthy])
-    return truthy ? now - startRunning : undefined
-}
